+++ src/main/java/com/sy/backEndApiAkilina/controllers/MinistereController.java
@@ -2,10 +2,7 @@ package com.sy.backEndApiAkilina.controllers;
 
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.sy.backEndApiAkilina.configuration.SaveImage;
-import com.sy.backEndApiAkilina.models.ERole;
-import com.sy.backEndApiAkilina.models.Ministere;
-import com.sy.backEndApiAkilina.models.Role;
-import com.sy.backEndApiAkilina.models.User;
+import com.sy.backEndApiAkilina.models.*;
 import com.sy.backEndApiAkilina.repository.MinistereRepository;
 import com.sy.backEndApiAkilina.repository.RoleRepository;
 import com.sy.backEndApiAkilina.repository.UserRepository;
@@ -13,18 +10,15 @@ import com.sy.backEndApiAkilina.security.services.MinistereService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
-import java.util.Optional;
 
 @RequestMapping("/api/ministere")
 @RestController
 @Api(value = "ministere", description = "MANIPULATION DES DONNEES DE LA TABLE MINISTERE")
-@CrossOrigin(origins = "http://localhost:4200", maxAge = 3600, allowCredentials="true")
-
+@CrossOrigin(origins = {"http://localhost:4200","http://localhost:8100"}, maxAge = 3600, allowCredentials="true")
 public class MinistereController {
     @Autowired
     MinistereService ministereService;
@@ -40,7 +34,7 @@ public class MinistereController {
     @PostMapping("/ajouter")
 
     public Object add(@RequestParam(value = "ministere") String minis,
-                         @RequestParam(value = "file", required = true) MultipartFile file) {
+                      @RequestParam(value = "file", required = true) MultipartFile file) {
         Role role = roleRepository.findByName(ERole.ROLE_ADMIN);
         User admin = userRepository.findByRoles(role);
         try {
@@ -56,6 +50,7 @@ public class MinistereController {
             return e.getMessage();
         }
     }
+
     @ApiOperation(value = "LIRE MINISTERE")
     @GetMapping("/lire")
     public List<Ministere> read() {
@@ -64,20 +59,31 @@ public class MinistereController {
 
     @ApiOperation(value = "Nombre de ministere")
     @GetMapping("/afficher_ministere_nombre")
-    public int readNombre() {return ministereService.read().size();
+    public int readNombre() {
+        return ministereService.read().size();
     }
 
 
     @ApiOperation(value = "LIRE MINISTERE Par libelle")
     @GetMapping("/lireParLibelle/{libelle}")
-    public Ministere readMinistereParLibelle(@PathVariable("libelle") String libelle) {
-        return ministereRepository.findByLibelle(libelle);
+    public Object readMinistereParLibelle(@PathVariable("libelle") String libelle) {
+        try {
+            return ministereRepository.findByLibelle(libelle);
+        }catch (Exception e){
+            return e.getMessage();
+        }
+
     }
 
     @ApiOperation(value = "LIRE MINISTERE Par ID")
     @GetMapping("/lireParId/{id_ministere}")
-    public Optional<Ministere> readMinistereParID(@PathVariable("id_ministere") Long id_ministere) {
-        return ministereRepository.findById(id_ministere);
+    public Object readMinistereParID(@PathVariable("id_ministere") Long id_ministere) {
+        try {
+            return ministereRepository.findById(id_ministere);
+        }catch (Exception e){
+            return e.getMessage();
+        }
+
     }
 
     @ApiOperation(value = "MODIFICATION DES DONNEES DE LA TABLE MINISTERE")
@@ -101,10 +107,19 @@ public class MinistereController {
     }
 
     @ApiOperation(value = "SUPPRESION DES DONNEE DANS LA TABLE MINISTERE")
-    @DeleteMapping("/supprimer/{id_ministere}")
+    @DeleteMapping("/supprimer/{id_ministere}/{id_user}")
 
-    public String delete(@PathVariable Long id_ministere) {
-        return ministereService.delete(id_ministere);
+    public String delete(@PathVariable Long id_ministere, @PathVariable Long id_user) {
+        try {
+            Ministere ministere = ministereRepository.findById(id_ministere).get();
+            if (ministere.getUser().getId_user() == id_user) {
+                return ministereService.delete(id_ministere);
+            } else {
+                return "vous n'etes pas autorisé à faire cette action";
+            }
+
+        } catch (Exception e) {
+            return e.getMessage();
+        }
     }
 }
-
